+++ consumers-spark/src/main/scala/it/agilelab/bigdata/wasp/consumers/spark/writers/SparkWriters.scala
@@ -13,5 +13,5 @@ trait SparkStructuredStreamingWriter {
 }
 
 trait SparkBatchWriter {
-	def write(data: DataFrame)
+	def write(data: DataFrame) : Unit
 }